For posterity...
An older version of the game. A lot of inefficiency, redundant copying, ugly structuring, just an all around uncomfortable block of code. The Arena was an older design that eventually became the Referee

Note the awkward hand off of a board to every player, multiple validations for valid moves, and the complex if/else that Levi was becoming.